+++ sdk/storage/azure-storage-blob-nio/src/main/java/com/azure/storage/blob/nio/NioBlobInputStream.java
@@ -208,4 +208,8 @@ public final class NioBlobInputStream extends InputStream {
     BlobInputStream getBlobInputStream() {
         return blobInputStream;
     }
+
+    Path getPath() {
+        return this.path;
+    }
 }