+++ sdk/storage/azure-storage-blob-nio/src/test/java/com/azure/storage/blob/nio/AzureFileSystemProviderTest.groovy
@@ -25,6 +25,7 @@ import java.nio.ByteBuffer
 import java.nio.charset.Charset
 import java.nio.file.AccessDeniedException
 import java.nio.file.AccessMode
+import java.nio.file.ClosedFileSystemException
 import java.nio.file.DirectoryNotEmptyException
 import java.nio.file.FileAlreadyExistsException
 import java.nio.file.FileSystem
@@ -334,7 +335,7 @@ class AzureFileSystemProviderTest extends APISpec {
         fs.provider().createDirectory(path)
 
         then:
-        thrown(IOException)
+        thrown(ClosedFileSystemException)
     }
 
     @Unroll
@@ -673,7 +674,7 @@ class AzureFileSystemProviderTest extends APISpec {
         fs.provider().copy(sourcePath, destPath, StandardCopyOption.COPY_ATTRIBUTES)
 
         then:
-        thrown(IOException)
+        thrown(ClosedFileSystemException)
 
         where:
         sourceClosed | _
@@ -775,7 +776,7 @@ class AzureFileSystemProviderTest extends APISpec {
         fs.provider().delete(path)
 
         then:
-        thrown(IOException)
+        thrown(ClosedFileSystemException)
     }
 
     def "DirectoryStream"() {
@@ -817,7 +818,7 @@ class AzureFileSystemProviderTest extends APISpec {
         fs.provider().newDirectoryStream(path, null)
 
         then:
-        thrown(IOException)
+        thrown(ClosedFileSystemException)
     }
 
     @Unroll
@@ -896,7 +897,7 @@ class AzureFileSystemProviderTest extends APISpec {
         fs.provider().newInputStream(path)
 
         then:
-        thrown(IOException)
+        thrown(ClosedFileSystemException)
     }
 
     def "OutputStream options default"() {
@@ -1068,7 +1069,7 @@ class AzureFileSystemProviderTest extends APISpec {
         fs.provider().newOutputStream(path)
 
         then:
-        thrown(IOException)
+        thrown(ClosedFileSystemException)
     }
 
     def "CheckAccess"() {
@@ -1214,7 +1215,7 @@ class AzureFileSystemProviderTest extends APISpec {
         fs.provider().checkAccess(path)
 
         then:
-        thrown(IOException)
+        thrown(ClosedFileSystemException)
     }
 
     @Unroll
@@ -1309,7 +1310,7 @@ class AzureFileSystemProviderTest extends APISpec {
         fs.provider().readAttributes(path, AzureBasicFileAttributes.class)
 
         then:
-        thrown(IOException)
+        thrown(ClosedFileSystemException)
     }
 
     @Unroll
@@ -1418,7 +1419,7 @@ class AzureFileSystemProviderTest extends APISpec {
         fs.provider().readAttributes(path, "basic:*")
 
         then:
-        thrown(IOException)
+        thrown(ClosedFileSystemException)
     }
 
     @Unroll
@@ -1578,7 +1579,7 @@ class AzureFileSystemProviderTest extends APISpec {
         fs.provider().setAttribute(path, "azureBlob:blobHttpHeaders", new BlobHttpHeaders())
 
         then:
-        thrown(IOException)
+        thrown(ClosedFileSystemException)
     }
 
     def basicSetupForCopyTest(FileSystem fs) {